gas optimization for manager

+++ contracts/interfaces/IOrderManager.sol
@@ -18,7 +18,7 @@ interface IOrderManager {
 
     function processLimitOrder(
         uint256 _tokenId, uint256 _batchId, uint256 _serviceFee, uint256 _monitorFee
-    ) external returns (uint256, uint256);
+    ) external returns (uint128, uint128);
 
     function canProcess(uint256 _tokenId, uint256 gasPrice) external view returns (bool, uint256, uint256);
